examples: move driver metrics out of the basic example
`basic.rs` is where README.md and the book send newcomers first, but a
third of it was a dump of every counter, latency percentile and
throughput rate the driver exposes. That is a topic of its own, and it
drowned out the basics the example is supposed to teach.

The metrics code moves verbatim into its own example, which - unlike
`basic.rs` - first issues enough traffic for the numbers to be
meaningful, and exercises both the unpaged and the automatically paged
request counters.
