+++ examples/basic.rs
@@ -89,50 +89,6 @@ async fn main() -> Result<()> {
         println!("a, b, c: {a}, {b}, {c}");
     }
 
-    let metrics = session.get_metrics();
-    println!(
-        "Unpaged queries requested: {}",
-        metrics.get_requests_unpaged_num()
-    );
-    println!(
-        "Automatically paged queries requested: {}",
-        metrics.get_requests_automatically_paged_num()
-    );
-    println!(
-        "Errors occurred in unpaged requests: {}",
-        metrics.get_errors_unpaged_num()
-    );
-    println!(
-        "Errors occurred in automatically paged requests: {}",
-        metrics.get_errors_automatically_paged_num()
-    );
-    println!("Average latency: {}", metrics.get_latency_avg_ms()?);
-    println!(
-        "99.9 latency percentile: {}",
-        metrics.get_latency_percentile_ms(99.9)?
-    );
-
-    let snapshot = metrics.get_snapshot()?;
-    println!("Min: {}", snapshot.min);
-    println!("Max: {}", snapshot.max);
-    println!("Mean: {}", snapshot.mean);
-    println!("Standard deviation: {}", snapshot.stddev);
-    println!("Median: {}", snapshot.median);
-    println!("75th percentile: {}", snapshot.percentile_75);
-    println!("95th percentile: {}", snapshot.percentile_95);
-    println!("98th percentile: {}", snapshot.percentile_98);
-    println!("99th percentile: {}", snapshot.percentile_99);
-    println!("99.9th percentile: {}", snapshot.percentile_99_9);
-
-    println!("Mean rate: {}", metrics.get_mean_rate());
-    println!("One minute rate: {}", metrics.get_one_minute_rate());
-    println!("Five minute rate: {}", metrics.get_five_minute_rate());
-    println!("Fifteen minute rate: {}", metrics.get_fifteen_minute_rate());
-
-    println!("Total connections: {}", metrics.get_total_connections());
-    println!("Connection timeouts: {}", metrics.get_connection_timeouts());
-    println!("Requests timeouts: {}", metrics.get_request_timeouts());
-
     println!("Ok.");
 
     Ok(())
